refactor: reorganize upload pages
