+++ resources/js/pages/upload/componentes/UploadDataTable.tsx
